+++ libs/logreg_classifier.cpp
@@ -19,7 +19,7 @@ LogregClassifier::LogregClassifier(const coef_t& coef) : coef_ {coef} {
   assert(!coef_.empty());
 }
 
-float LogregClassifier::predict_proba(const features_t& feat) const {
+double LogregClassifier::predict_proba(const features_t& feat) const {
   auto z = std::inner_product(feat.begin(),
                               feat.end(),
                               ++coef_.begin(),
